# api-spec.yml
openapi: 3.0.0
info:
  title: Views and Reactions API
  version: 1.0.0
  description: API to manage views and reactions for articles and snippets.

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: base64
    
    HTTPSOnly:
      type: http
      scheme: https

security:
  - BearerAuth: []

paths:
  /views:
    get:
      summary: Get view count
      description: Retrieve the number of views for the specified path.
      security:
        - BearerAuth: []
      parameters:
        - name: path
          in: query
          required: true
          description: The path of the content.
          schema:
            type: string
      responses:
        '200':
          description: View count retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  views:
                    type: integer
        '401':
          description: Unauthorized - Invalid or missing authentication token

    put:
      summary: Increment view count
      description: Increment the view count for the specified path.
      security:
        - BearerAuth: []
      parameters:
        - name: path
          in: query
          required: true
          description: The path of the content.
          schema:
            type: string
      responses:
        '200':
          description: View count incremented successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  views:
                    type: integer
        '401':
          description: Unauthorized - Invalid or missing authentication token

  /reactions:
    get:
      summary: Get reaction counts
      description: Retrieve the count for each reaction for the specified path.
      security:
        - BearerAuth: []
      parameters:
        - name: path
          in: query
          required: true
          description: The path of the content.
          schema:
            type: string
      responses:
        '200':
          description: Reaction counts retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  love:
                    type: integer
                  like:
                    type: integer
                  mindblown:
                    type: integer
                  puzzling:
                    type: integer
        '401':
          description: Unauthorized - Invalid or missing authentication token

    put:
      summary: Increment reaction count
      description: Increment the count for a specific reaction in the specified path.
      security:
        - BearerAuth: []
      parameters:
        - name: path
          in: query
          required: true
          description: The path of the content.
          schema:
            type: string
        - name: reaction
          in: query
          required: true
          description: The reaction type.
          schema:
            type: string
            enum: [love, like, mindblown, puzzling]
      responses:
        '200':
          description: Reaction count incremented successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  love:
                    type: integer
                  like:
                    type: integer
                  mindblown:
                    type: integer
                  puzzling:
                    type: integer
        '401':
          description: Unauthorized - Invalid or missing authentication token